search: Use size_t for swap count parameter
Don't use a random collection of 'int' and 'long' types.

+++ newlib/libc/search/qsort.c
@@ -79,7 +79,7 @@ typedef int		 cmp_t(const void *, const void *, void *);
 typedef int		 cmp_t(const void *, const void *);
 #endif
 static inline char	*med3 (char *, char *, char *, cmp_t *, void *);
-static inline void	 swapfunc (char *, char *, int, int);
+static inline void	 swapfunc (char *, char *, size_t, int);
 
 #define min(a, b)	(a) < (b) ? a : b
 
@@ -87,7 +87,7 @@ static inline void	 swapfunc (char *, char *, int, int);
  * Qsort routine from Bentley & McIlroy's "Engineering a Sort Function".
  */
 #define swapcode(TYPE, parmi, parmj, n) { 		\
-	long i = (n) / sizeof (TYPE); 			\
+	size_t i = (n) / sizeof (TYPE); 			\
 	TYPE *pi = (TYPE *) (parmi); 		\
 	TYPE *pj = (TYPE *) (parmj); 		\
 	do { 						\
@@ -103,7 +103,7 @@ static inline void	 swapfunc (char *, char *, int, int);
 static inline void
 swapfunc (char *a,
 	char *b,
-	int n,
+	size_t n,
 	int swaptype)
 {
 	if(swaptype <= 1)
